Minor Refactoring
WithSuper: deleted;
Reflection and Utilities: moved to package core;
Extractors: extractField: moved and refactored;
