--- src/test/scala/com/phasmidsoftware/core/WithSuperTest.scala
@@ -1,29 +1,0 @@
-package com.phasmidsoftware.core
-
-import org.scalatest.flatspec.AnyFlatSpec
-
-class WithSuperTest extends AnyFlatSpec {
-
-    behavior of "WithSuperTest"
-
-    it should "productElementName" in {
-
-    }
-
-    it should "productPrefix" in {
-
-    }
-
-    it should "productElementNames" in {
-
-    }
-
-    it should "superObject" in {
-
-    }
-
-    it should "productIterator" in {
-
-    }
-
-}
